move new router to backendBridgeBundle
