+++ EMSBackendBridgeBundle/DependencyInjection/EMSBackendBridgeExtension.php
@@ -29,6 +29,7 @@ class EMSBackendBridgeExtension extends Extension
     {
         $loader = new XmlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
         $loader->load('services.xml');
+        $loader->load('routing.xml');
 
         $configuration = new Configuration();
         $config = $this->processConfiguration($configuration, $configs);
