--- EMSRoutingBundle/Resources/config/services.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<container xmlns="http://symfony.com/schema/dic/services"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://symfony.com/schema/dic/services
-        http://symfony.com/schema/dic/services/services-1.0.xsd">
-
-    <services>
-        <defaults autowire="false" autoconfigure="false" public="false" />
-
-
-        <service id="EMS\ClientHelperBundle\EMSRoutingBundle\Routing\Router">
-            <argument><!-- clientRequest injected by configPass --></argument>
-            <argument type="service" id="router"/>
-            <argument type="service" id="templating"/>
-            <tag name="controller.service_arguments" />
-        </service>
-        <service id="EMS\ClientHelperBundle\EMSRoutingBundle\Routing\RouteLoader">
-            <argument type="string">%kernel.project_dir%</argument>
-            <tag name="routing.loader" />
-        </service>
-        
-        <prototype namespace="EMS\ClientHelperBundle\EMSRoutingBundle\" resource="../../../EMSRoutingBundle/*" exclude="../../../EMSRoutingBundle/Routing" />
-    </services>
-</container>
